+++ example/system/amp/openamp_for_linux/sdkconfig.h
@@ -17,26 +17,31 @@
 
 /* Arm architecture configuration */
 
-#define CONFIG_ARCH_ARMV8_AARCH64
-/* CONFIG_ARCH_ARMV8_AARCH32 is not set */
+/* CONFIG_ARCH_ARMV8_AARCH64 is not set */
+#define CONFIG_ARCH_ARMV8_AARCH32
 
 /* Compiler configuration */
 
 #define CONFIG_ARM_GCC_SELECT
 /* CONFIG_ARM_CLANG_SELECT is not set */
 #define CONFIG_TOOLCHAIN_NAME "gcc"
-#define CONFIG_TARGET_ARMV8_AARCH64
-#define CONFIG_ARCH_EXECUTION_STATE "aarch64"
-#define CONFIG_ARM_NEON
-#define CONFIG_ARM_CRC
-#define CONFIG_ARM_CRYPTO
-#define CONFIG_ARM_FLOAT_POINT
-/* CONFIG_GCC_CODE_MODEL_TINY is not set */
-#define CONFIG_GCC_CODE_MODEL_SMALL
-/* CONFIG_GCC_CODE_MODEL_LARGE is not set */
+#define CONFIG_TARGET_ARMV8_AARCH32
+#define CONFIG_ARCH_EXECUTION_STATE "aarch32"
+
+/* Fpu configuration */
+
+#define CONFIG_CRYPTO_NEON_FP_ARMV8
+/* CONFIG_VFPV4 is not set */
+/* CONFIG_VFPV4_D16 is not set */
+/* CONFIG_VFPV3 is not set */
+/* CONFIG_VFPV3_D16 is not set */
+#define CONFIG_ARM_MFPU "crypto-neon-fp-armv8"
+#define CONFIG_MFLOAT_ABI_HARD
+/* CONFIG_MFLOAT_ABI_SOFTFP is not set */
+#define CONFIG_ARM_MFLOAT_ABI "hard"
+/* end of Fpu configuration */
 /* end of Compiler configuration */
-/* CONFIG_BOOT_WITH_FLUSH_CACHE is not set */
-/* CONFIG_MMU_DEBUG_PRINTS is not set */
+#define CONFIG_USE_AARCH64_L1_TO_AARCH32
 /* end of Arm architecture configuration */
 
 /* multi-core system deployment framework */
@@ -44,20 +49,22 @@
 /* CONFIG_USE_MSDF is not set */
 /* end of multi-core system deployment framework */
 #define CONFIG_MMU_PAGE_SIZE 0x1000
-#define CONFIG_MAX_XLAT_TABLES 256
+#define CONFIG_FMMU_NUM_L2_TABLES 256
+/* CONFIG_ENABLE_GIC_ITS is not set */
 /* end of Arch configuration */
 
 /* Soc configuration */
 
-#define CONFIG_TARGET_PHYTIUMPI
-/* CONFIG_TARGET_E2000Q is not set */
+/* CONFIG_TARGET_PHYTIUMPI is not set */
+#define CONFIG_TARGET_E2000Q
 /* CONFIG_TARGET_E2000D is not set */
 /* CONFIG_TARGET_E2000S is not set */
 /* CONFIG_TARGET_FT2004 is not set */
 /* CONFIG_TARGET_D2000 is not set */
 /* CONFIG_TARGET_PD2308 is not set */
 /* CONFIG_TARGET_QEMU_VIRT is not set */
-#define CONFIG_SOC_NAME "phytiumpi"
+#define CONFIG_SOC_NAME "e2000"
+#define CONFIG_TARGET_TYPE_NAME "q"
 #define CONFIG_SOC_CORE_NUM 4
 #define CONFIG_F32BIT_MEMORY_ADDRESS 0x80000000
 #define CONFIG_F32BIT_MEMORY_LENGTH 0x80000000
@@ -71,7 +78,7 @@
 
 /* Board Configuration */
 
-#define CONFIG_BOARD_NAME "firefly"
+#define CONFIG_BOARD_NAME "demo"
 /* CONFIG_USE_SPI_IOPAD is not set */
 /* CONFIG_USE_GPIO_IOPAD is not set */
 /* CONFIG_USE_CAN_IOPAD is not set */
@@ -81,7 +88,7 @@
 /* CONFIG_USE_TACHO_IOPAD is not set */
 /* CONFIG_USE_UART_IOPAD is not set */
 /* CONFIG_USE_THIRD_PARTY_IOPAD is not set */
-#define CONFIG_FIREFLY_DEMO_BOARD
+#define CONFIG_E2000Q_DEMO_BOARD
 
 /* IO mux configuration when board start up */
 
@@ -191,7 +198,12 @@
 #define CONFIG_IMAGE_LOAD_ADDRESS 0xb0100000
 #define CONFIG_IMAGE_MAX_LENGTH 0x2000000
 #define CONFIG_HEAP_SIZE 1
-#define CONFIG_STACK_SIZE 0x400
+#define CONFIG_SVC_STACK_SIZE 0x1000
+#define CONFIG_SYS_STACK_SIZE 0x1000
+#define CONFIG_IRQ_STACK_SIZE 0x1000
+#define CONFIG_ABORT_STACK_SIZE 0x1000
+#define CONFIG_FIQ_STACK_SIZE 0x1000
+#define CONFIG_UNDEF_STACK_SIZE 0x1000
 /* end of Linker Options */
 /* end of Build setup */